implemented support for multiple PIC drivers

+++ src/kernel/arch/i686/irq.h
@@ -5,4 +5,4 @@
 typedef void (*IRQHandler)(Registers* regs);
 
 void i686_IRQ_Initialize();
-void i686_IRQ_RegisterHandler(uint8_t irq, IRQHandler handler);
+void i686_IRQ_RegisterHandler(int irq, IRQHandler handler);
